Configs don't need `CallRemoteConfigOperation` anymore
Needs tidyup but we now run coroutines needed for configs on a single MonoBehaviour.
Moved some code from SdkOperations to BackendOperations
